[Patch v4.8.13] Merge strategy modules
